+++ src/components/Share/ShareContent.jsx
@@ -1,4 +1,4 @@
-import Profile  from "../Profile";
+import Profile  from "./Profile";
 import Content from "../Content/Content";
 import {useEffect, useState} from 'react';
 import { API_PATH_SAMPLE_FOLDER } from "../../services/api-path";
